moving init for variable pages
