+++ Core/Inc/top_board/oled/variable_pages/time_page.h
@@ -3,7 +3,9 @@
 
 #include "pages.h"
 #include "robot.h"
+#include <string.h>
 
+void time_page_init(page_struct* parent);
 void time_page_update(page_struct *page);
 
 #endif /* INC_TIME_PAGE_H_ */
