+++ Core/Src/top_board/oled/variable_pages/encoder_page.c
@@ -1,5 +1,20 @@
 #include "encoder_page.h"
 
+struct page_struct encoder_values;
+
+void encode_page_init(page_struct *parent) {
+    pages_set_default_values(&encoder_values);
+    encoder_values.id = 24;
+    strcpy(encoder_values.page_name, "Encoder");
+    strcpy(encoder_values.line0, "LF:0");
+    strcpy(encoder_values.line1, "RF:0");
+    strcpy(encoder_values.line2, "LB:0");
+    strcpy(encoder_values.line3, "RB:0");
+    encoder_values.parent = parent;
+    encoder_values.has_variables = true;
+    add_child_to_parent(&encoder_values);
+}
+
 void encoder_page_update(page_struct *page){
 
     char temp[MAX_STRING_LENGTH];
